requirements.txt
flask>=3.0.0
flask-socketio>=5.3.0
python-dotenv>=1.0.0
telethon>=1.33.0
asyncio>=3.4.0
pytz>=2024.0
aiohttp>=3.9.0
aiofiles>=23.2.0
werkzeug>=3.0.0
flask-login>=0.6.0
pydantic>=2.7.0
watchdog>=4.0.0
eventlet>=0.35.0
dnspython>=2.6.0
marshmallow>=3.21.0
pillow>=10.3.0
python-magic>=0.4.0
jsonschema>=4.21.0
gevent>=24.2.0

# flask
# flask-socketio
# python-dotenv
# telethon
# asyncio
# pytz
# aiohttp
# aiofiles
# werkzeug
# flask-login
# pydantic
# watchdog
# eventlet
# marshmallow
# pillow
# python-magic
# jsonschema
# gevent


# flask==3.0.2
# flask-socketio==5.3.6
# python-dotenv==1.0.1
# telethon==1.33.1
# asyncio==3.4.3
# pytz==2024.1
# aiohttp==3.9.5
# aiofiles==23.2.1
# werkzeug==3.0.2
# flask-login==0.6.3
# pydantic==2.7.0
# watchdog==4.0.0
# eventlet==0.35.2
# marshmallow==3.21.1
# pillow==10.3.0 
# python-magic==0.4.27
# jsonschema==4.21.1
# gevent==24.2.1